+++ lib/conditional_trigger.dart
@@ -1,5 +1,4 @@
 library conditional_trigger;
 
 export 'src/conditional_trigger.dart';
-export 'src/models/conditional_mocks.dart';
 export 'src/models/conditional_state.dart';