+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.3
+
+* Add `lastState` to `ConditionalTrigger` to get the last state of the `check()`.
+* Add `dispose` method as an optional to free the memory.
+* Able to clear all last states by using `ConditionalTrigger.clearAllStates`.
+* Update README.
+
 ## 0.0.2
 
 * Initial release.